refactor: extract routes into separate modules
